Add internal metadata helper header
